requirements.txt
streamlit
python-dotenv
openai
requests
Pillow
gTTS
konlpy
moviepy
